refactor: 경로 찾기 알고리즘 로직 수정 및 테스트 메소드 위치 변경
